backend/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
